refactor(workflow-tasks): move k8s e2e remote-script builder into library
